Application/requirements.txt
scikit-surprise
scikit-learn
pandas
numpy
matplotlib
seaborn